+++ org/xbill/DNS/NSRecord.java
@@ -20,7 +20,7 @@ private
 NSRecord() {}
 
 private
-NSRecord(Name name, int dclass, int ttl) {
+NSRecord(Name name, int dclass, long ttl) {
 	super(name, Type.NS, dclass, ttl);
 }
 
@@ -34,12 +34,12 @@ getMember() {
  * @param target The name server for the given domain
  */
 public
-NSRecord(Name name, int dclass, int ttl, Name target) {
+NSRecord(Name name, int dclass, long ttl, Name target) {
 	super(name, Type.NS, dclass, ttl, target);
 }
 
 Record
-rrFromWire(Name name, int type, int dclass, int ttl, int length,
+rrFromWire(Name name, int type, int dclass, long ttl, int length,
 	   DataByteInputStream in)
 throws IOException
 {
@@ -47,7 +47,7 @@ throws IOException
 }
 
 Record
-rdataFromString(Name name, int dclass, int ttl, Tokenizer st, Name origin)
+rdataFromString(Name name, int dclass, long ttl, Tokenizer st, Name origin)
 throws IOException
 {
 	return rdataFromString(new NSRecord(name, dclass, ttl), st, origin);
